# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bank Mondial API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /sub-accounts:
    get:
      operationId: listSubAccounts
      tags: [sub-accounts]
      summary: List sub-accounts
      responses:
        "200":
          description: List of sub-accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubAccount"
    post:
      operationId: createSubAccount
      tags: [sub-accounts]
      summary: Create sub-account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubAccount"
  /sub-accounts/{id}:
    get:
      operationId: getSubAccount
      tags: [sub-accounts]
      summary: Get sub-account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sub-account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubAccount"
    patch:
      operationId: updateSubAccount
      tags: [sub-accounts]
      summary: Update sub-account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubAccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubAccount"
    delete:
      operationId: deleteSubAccount
      tags: [sub-accounts]
      summary: Delete sub-account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, blocked, pending]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}/block:
    patch:
      operationId: adminBlockUser
      tags: [admin]
      summary: Admin - block/unblock user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/balance:
    patch:
      operationId: adminUpdateBalance
      tags: [admin]
      summary: Admin - update user balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/transfers:
    get:
      operationId: adminListTransfers
      tags: [admin]
      summary: Admin - list all transfers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transfers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferList"
...
  /admin/kyc/{id}/review:
    patch:
      operationId: adminReviewKyc
      tags: [admin]
      summary: Admin - approve/reject KYC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycReview"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kyc"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
      required:
        - fullName
        - email
        - phone
        - country
        - password
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        balance:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [active, blocked, pending]
        role:
          type: string
          enum: [user, admin]
        referralCode:
          type: string
        kycStatus:
          type: string
          enum: [none, pending, verified, rejected]
        iban:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clientId
        - fullName
        - email
        - phone
        - country
        - balance
        - currency
        - status
        - role
        - referralCode
        - kycStatus
        - createdAt
...
    Transfer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        token:
          type: string
        beneficiaryName:
          type: string
        amount:
          type: number
        currency:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, cancelled, expired]
        accessType:
          type: string
          enum: [public, private, limited]
        expiresAt:
          type: ["string", "null"]
        reference:
          type: string
        createdAt:
          type: string
        confirmedAt:
          type: ["string", "null"]
        linkUrl:
          type: string
      required:
        - id
        - userId
        - token
        - beneficiaryName
        - amount
        - currency
        - status
        - accessType
        - reference
        - createdAt
        - linkUrl
    TransferList:
      type: object
      properties:
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - transfers
        - total
        - page
        - limit
...
    SubAccount:
      type: object
      properties:
        id:
          type: integer
        parentUserId:
          type: integer
        fullName:
          type: string
        email:
          type: string
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
      required:
        - id
        - parentUserId
        - fullName
        - email
        - permissions
        - status
        - createdAt
    SubAccountInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        permissions:
          type: array
          items:
            type: string
      required:
        - fullName
        - email
        - password
        - permissions
    SubAccountUpdate:
      type: object
      properties:
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, suspended]
...
    Kyc:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        documentType:
          type: string
          enum: [id_card, passport, driving_license]
        documentNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, verified, rejected]
        rejectionReason:
          type: ["string", "null"]
        submittedAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
      required:
        - id
        - userId
        - documentType
        - status
        - submittedAt
...
    KycReview:
      type: object
      properties:
        status:
          type: string
          enum: [verified, rejected]
        rejectionReason:
          type: ["string", "null"]
      required:
        - status
...
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    BlockInput:
      type: object
      properties:
        blocked:
          type: boolean
      required:
        - blocked
    BalanceUpdate:
      type: object
      properties:
        balance:
          type: number
        reason:
          type: string
      required:
        - balance
        - reason
...