# lib/api-zod/src/generated/types/subAccountUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bank Mondial API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubAccountUpdateStatus } from './subAccountUpdateStatus';

export interface SubAccountUpdate {
  permissions?: string[];
  status?: SubAccountUpdateStatus;
}