# lib/api-zod/src/generated/types/transferList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bank Mondial API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Transfer } from './transfer';

export interface TransferList {
  transfers: Transfer[];
  total: number;
  page: number;
  limit: number;
}