# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bank Mondial API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycStatus } from './userKycStatus';
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  clientId: string;
  fullName: string;
  email: string;
  phone: string;
  country: string;
  balance: number;
  currency: string;
  status: UserStatus;
  role: UserRole;
  referralCode: string;
  kycStatus: UserKycStatus;
  /** @nullable */
  iban?: string | null;
  createdAt: string;
}